requirements.txt
speedtest-cli==2.1.1
twitter==1.18.0
